// src/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      subscriptions: {
        Row: {
          billing_anchor: number | null
          cancel_at_period_end: boolean
          created_at: string | null
          ends_at: string | null
          id: number
          renews_at: string | null
          status: Database["public"]["Enums"]["subscription_status"]
          trial_ends_at: string | null
          trial_starts_at: string | null
          update_payment_method_url: string | null
          variant_id: number
        }
        Insert: {
          billing_anchor?: number | null
          cancel_at_period_end: boolean
          created_at?: string | null
          ends_at?: string | null
          id: number
          renews_at?: string | null
          status: Database["public"]["Enums"]["subscription_status"]
          trial_ends_at?: string | null
          trial_starts_at?: string | null
          update_payment_method_url?: string | null
          variant_id: number
        }
        Update: {
          billing_anchor?: number | null
          cancel_at_period_end?: boolean
          created_at?: string | null
          ends_at?: string | null
          id?: number
          renews_at?: string | null
          status?: Database["public"]["Enums"]["subscription_status"]
          trial_ends_at?: string | null
          trial_starts_at?: string | null
          update_payment_method_url?: string | null
          variant_id?: number
        }
        Relationships: []
      }
// ...
    }
// ...
  }
// ...
}

type PublicSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never